utils naming updated
