+++ src/components/spreadsheet-import/index.tsx
@@ -3,8 +3,8 @@ import { Fragment, useState } from "react";
 import { ReactSpreadsheetImport } from "@wavepoint/react-spreadsheet-import";
 import { Result } from "@wavepoint/react-spreadsheet-import/types/types";
 import { StyledBackdrop } from "./index.styled";
-import { getImportFields } from "utils/import-key-mappings";
 import { useCoreModuleNavigation, useNotificationsService } from "@hooks";
+import { getImportFields } from "utils/import-file-helper";
 
 interface csvImportPorps {
   isOpen: boolean;
